Shrink ShiftedValueIndex from 16 bytes to 8
The shift amount in a ShiftedValueIndex is always less than 64, yet it
was stored as a usize (8 bytes). With alignment padding that made the
whole struct 16 bytes:

    value_index: u32 | shift_variant | amount: usize   -> 16 bytes

Constraint systems hold millions of these, and the prover reads them on
its hot path while evaluating operands. Storing the amount as a u8 drops
the struct to a single 8-byte word, halving the memory and doubling the
cache density of that traversal:

    value_index: u32 | shift_variant | amount: u8       ->  8 bytes

The public constructors still take amount: usize, so no call site
changes; they narrow to u8 internally. The serialized wire format is
kept identical (usize serializes as a u32 in this repo, so the amount is
still written as a u32 and narrowed on read after the range check), which
means no serialization-version bump and no reference-binary regeneration.

+++ crates/core/src/constraint_system/system.rs
@@ -100,7 +100,7 @@ impl ConstraintSystem {
 						constraint_type,
 						constraint_index,
 						operand_name,
-						shift_amount: term.amount,
+						shift_amount: term.amount as usize,
 					});
 				}
 				// Check if the value index is out of bounds.
